In the middle of adding a "submit" button.
I now have a button that can give a file location to Flask, which then executes the code.
Any errors are caught, but I need to send them back to the user somehow.
The code is currently written inline. I'll need to collect all this functionality into a reuseable class.